+++ src/plugins/Attr.js
@@ -89,7 +89,7 @@ export const AttrPlugin = {
    * Plugin version
    * @type {string}
    */
-  version: "1.1.0",
+  version: "1.1.1",
 
   /**
    * Plugin description